Organized it in the new component
